:art: Extract IndexBanner
