+++ src/app/commons/IndexBanner/IndexBanner.scss
